% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigobenchmark.R
\name{autoplot.bigobenchmark}
\alias{autoplot.bigobenchmark}
\title{Autoplot for bigobenchmark object}
\usage{
\method{autoplot}{bigobenchmark}(object)
}
\arguments{
\item{object}{}
}
\value{
A ggplot2 plot
}
\description{
Autoplot for bigobenchmark object
}
\examples{
# Create plot for benchmarks
library(ggplot2)
bench <- bigobenchmark(cumsum(1:n), cumprod(1:n), args=seq(from=1000, to=10000, length.out = 50))
autoplot(bench)

}
\author{
Andrew Prokhorenkov
}
