% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigobenchmark.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot.bigobenchmark}
\title{Autoplot for bigobenchmark object}
\usage{
autoplot.bigobenchmark(object)
}
\description{
Autoplot for bigobenchmark object
}
\examples{
# Create plot for benchmarks
library(ggplot2)
bench <- bigobenchmark(1:n, for(i in 1:n) for(i in 1:n) 1:n, args=seq(from=1, to=100, length.out = 50))
autoplot(bench)

}
\author{
Andrew Prokhorenkov
}
