% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigobenchmark.R
\name{bigobenchmark}
\alias{bigobenchmark}
\title{bigobenchmark runner}
\usage{
bigobenchmark(..., args = NULL, list = NULL, times = 100L, check = NULL,
  control = list())
}
\arguments{
\item{...}{Expressions to benchmark. Passed to \code{microbenchmark}.}

\item{args}{List of arguments to run functions on.}

\item{list}{List of unevaluated expression to benchmark. Passed to \code{microbenchmark}.}

\item{times}{Number of times to evaluate the expression. Passed to \code{microbenchmark}.}

\item{check}{Function to check if the expressions are equal. By default \code{NULL} which omits the check. Passed to \code{microbenchmark}.}

\item{control}{List of control arguments. See Details section for \code{microbenchmark}. Passed to \code{microbenchmark}.}
}
\value{
An object of class \code{bigobenchmark}.
}
\description{
bigobenchmark runner
}
\examples{
# Measure difference between linear and quadratic functions
bench <- bigobenchmark(1:n, for(i in 1:n) for(i in 1:n) 1:n, args=seq(from=1, to=100, length.out = 50))
bench

}
\author{
Andrew Prokhorenkov
}
